requirements.txt
# Core dependencies
requests>=2.31.0
slack-sdk>=3.23.0
PyGitHub>=1.59.0
facebook-sdk>=3.1.0

# AI and ML
langchain>=0.1.0
langchain-community>=0.3.0
sentence-transformers>=2.2.2
openai>=1.3.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
google-generativeai>=0.3.0

# Database and Storage
psycopg2-binary>=2.9.7
pyarrow>=12.0.0
pandas>=2.0.0
boto3>=1.34.0
PyYAML>=6.0.0

# Web Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
redis>=4.0.0
psutil>=5.9.0
schedule>=1.2.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
